check roles,bus_stop map,delete users,bus_stops

+++ migrations/20220613133955_bus_stops.sql
@@ -2,8 +2,8 @@
 CREATE TABLE bus_stops (
     id int AUTO_INCREMENT,
     name varchar(200) NOT NULL,
-    loc_lat DECIMAL(8,6) CHECK(loc_lat IS NULL OR (loc_lat >= -90 AND loc_lat <= 90)),
-    loc_long DECIMAL(9,6) CHECK(loc_lat IS NULL OR (loc_long >= -180 AND loc_long <= 180)),
+    loc_lat FLOAT(8,6) CHECK(loc_lat IS NULL OR (loc_lat >= -90 AND loc_lat <= 90)),
+    loc_long FLOAT(9,6) CHECK(loc_lat IS NULL OR (loc_long >= -180 AND loc_long <= 180)),
     address TEXT NOT NULL, 
     PRIMARY KEY(id)
 );
